Download after

+++ src/core/main/services/DownloadService.ts
@@ -6,37 +6,47 @@ import { Media } from "intus-database/WatermelonDB/models/Media/Media";
 export const MAX_TRIES = 3;
 export const MEDIAS_DIR = `${FileSystem.documentDirectory}/medias`;
 
-export const mediaDownloadHandler = async (media: Media): Promise<Media> => {
-	const {
-		data: { temp_url: DOWNLOAD_URL },
-	} = await mediaDownloadTempURLRequest(media.filename);
-	const DOWNLOAD_PATH = await makeDownloadPath(media.filename);
-
-	let tries = 1;
-	let timeout = 1000 * tries;
-
-	while (tries <= MAX_TRIES) {
-		try {
-			tries = tries + 1;
-			timeout = 1000 * tries;
-
-			await awaitableTimeout(timeout);
-			const download = await FileSystem.downloadAsync(DOWNLOAD_URL, DOWNLOAD_PATH);
-
-			if (!isStatusCodeOk(download)) {
-				// We try to delete the file if the status code is not 200
-				await deleteFile(DOWNLOAD_PATH);
-			} else {
-				await media.setDownloadedPath(download.uri);
-				return media;
-			}
-		} catch (e) {
-			console.log(`Could not download, retrying after ${timeout}`);
-		}
+export const mediaDownloadHandler = async (
+	media: Media
+): Promise<{ media: Media; downloadedPath: string }> => {
+	const fileExists = await mediaExists(media.filename);
+	if (fileExists) {
+		return { media, downloadedPath: fileExists.path };
 	}
 
-	// throw here works as reject() call inside a Promise.
-	throw new DownloadFailedError(media.media_id, MAX_TRIES);
+	try {
+		const {
+			data: { temp_url: DOWNLOAD_URL },
+		} = await mediaDownloadTempURLRequest(media.filename);
+		const DOWNLOAD_PATH = await makeDownloadPath(media.filename);
+
+		let tries = 1;
+		let timeout = 1000 * tries;
+
+		while (tries <= MAX_TRIES) {
+			try {
+				tries = tries + 1;
+				timeout = 1000 * tries;
+
+				await awaitableTimeout(timeout);
+				const download = await FileSystem.downloadAsync(DOWNLOAD_URL, DOWNLOAD_PATH);
+
+				if (!isStatusCodeOk(download)) {
+					// We try to delete the file if the status code is not 200
+					await deleteFile(DOWNLOAD_PATH);
+				} else {
+					return { media, downloadedPath: download.uri };
+				}
+			} catch (e) {
+				console.log(`Could not download, retrying after ${timeout}`);
+			}
+		}
+
+		// throw here works as reject() call inside a Promise.
+		throw new DownloadFailedError(media.media_id, MAX_TRIES);
+	} catch (e) {
+		throw e;
+	}
 };
 
 const isStatusCodeOk = (downloadResult: FileSystem.FileSystemDownloadResult | undefined) => {
